+++ example.js
@@ -2,12 +2,12 @@ var RaumfeldManager = require('node-raumfeld');
 
 var manager = new RaumfeldManager();
 
-// starts discovering devices asynchronously
+// starts discovering renderers asynchronously
 manager.discover();
 
-// fetch a device and do something with it. Method calls on a device return a promise object
-var device = manager.getDevice("Kitchen");
+// fetch a renderer and do something with it. Method calls on a renderer return a promise object
+var renderer = manager.getRenderer("Kitchen");
 
-device.getVolume().then(function(value) {
-    console.log("Device volume is " + value);
+renderer.getVolume().then(function(value) {
+    console.log("Renderer volume is " + value);
 });