+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - added virtual device to be able to handle room/zone capabilities
 - added getMediaInfo to virtual device
 - updated upnp client dependency
+- renamed device to renderer
 
 ## 0.1.0 (2016-03-14)
 - added functionality for mute control, play, pause and stop